don't have separate files when we don't need to

+++ train.py
@@ -10,37 +10,7 @@ import shutil, os, logging
 import gym
 import numpy as np
 from doom_py import ScreenResolution
-
-from skimage.color import rgb2gray
-from skimage.transform import resize
-
-
-class ObFilter(object):
-    def __init__(self, new_width, new_height):
-        self.w = new_width
-        self.h = new_height
-        self.f = Flatten()
-
-    def __call__(self, ob):
-        out = resize(rgb2gray(ob), (self.h, self.w))
-        return self.f(out)
-
-    def output_shape(self, input_shape):
-        return (self.h * self.w,)
-
-class ActFilter(object):
-    def __init__(self, lookup):
-        self.lookup = lookup
-        self.n = len(self.lookup)
-
-    def __call__(self, act):
-        action_list = np.zeros(43) # Doom has 43 actions
-        action_list[self.lookup[act]] = 1
-        return action_list
-
-    def output_shape(self):
-        return self.n
-
+from filters import ObFilterCNN, ObFilterFF, ActFilter
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
@@ -55,9 +25,12 @@ if __name__ == "__main__":
     env_spec = env.spec
 
     mondir = args.outfile + ".dir"
-    if os.path.exists(mondir): shutil.rmtree(mondir)
-    os.mkdir(mondir)
-    env.monitor.start(mondir, video_callable=None if args.video else VIDEO_NEVER)
+    if args.load_snapshot:
+        env.monitor.start(mondir, video_callable=None if args.video else VIDEO_NEVER, resume=True)
+    else:
+        if os.path.exists(mondir): shutil.rmtree(mondir)
+        os.mkdir(mondir)
+        env.monitor.start(mondir, video_callable=None if args.video else VIDEO_NEVER)
     agent_ctor = get_agent_cls(args.agent)
     update_argument_parser(parser, agent_ctor.options)
     args = parser.parse_args()
@@ -69,8 +42,13 @@ if __name__ == "__main__":
     # Setup environment and filters
     aa = env.__dict__['allowed_actions']
     action_mapping = {i: aa[i] for i in range(len(aa))}
-    of = ObFilter(20, 15)
     af = ActFilter(action_mapping)
+    if args.agent == 'modular_rl.agentzoo.TrpoAgent':
+        of = ObFilterFF(20, 15)
+    elif args.agent == 'modular_rl.agentzoo.TrpoAgentCNN':
+        of = ObFilterCNN(20, 15)
+    else:
+        raise ValueError('args.agent {} is invalid'.format(args.agent))
     envf = FilteredEnv(env, ob_filter=of, act_filter=af, skiprate=(3,7))
 
     print envf.observation_space, envf.action_space
